Add files via upload
This is a basic C++ project ,car rental system is a platform that manages the operations or the process between the admin and customer. In this project, I have added some data file(.txt) , which store data that is entered by the admin who manages the whole process. And the data entered by customer is also stored in the form of .txt file. So, the basic purpose is to make a unique platform for both the admin and the customer so that their demands (like the customer who wants to rent a car and the admin who is willingly waiting to rent them) can be fulfilled.